Make console driver more defensive
There were a couple cases where a buggy process could have influenced
it's own metatdata such that copying from the process slice to the
shared buffer used a negative index (if `write_remaining` was larger
than `slice.len()`), which would panic. This adds some defensive measures that don't fix
the processes behavior but contains the bug to the processes (gibberish
comes out, but the kernel and other processes are fine).