Add ALLOWRO system call
Adds a system call for allowing the kernel read-only access to process memory.  It also introduces a new marker type for AppPtr and AppSlice (SharedRO) that marks it read only. This allows us to separate implemenations for deref and deref_mut, as_ref and as_mut_ref on the basis of whether the underlying memory is RO or RW